사용자의 답 검증하는 코드를 User에 이동
